Last MAJ 20.04

+++ game/Kusinta/automaton/playerActions/Action.java
@@ -2,14 +2,14 @@ package playerActions;
 
 public abstract class Action implements IAction {
 	
-	public int percentage;
+	public float percentage;
 	
 	public Action() {
 		percentage = -1;
 	}
 	
 	public Action(int percentage) {
-		this.percentage = percentage;
+		this.percentage = (float) percentage;
 	}
 
 }
